Finish.

+++ demo-1/src/main/java/com/example/demo/entity/WatchEntity.java
@@ -1,18 +1,14 @@
 package com.example.demo.entity;
 
-import java.sql.Blob;
-import java.sql.SQLException;
-import java.util.Base64;
+import java.math.BigDecimal;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.persistence.Lob;
-import javax.persistence.Transient;
 
 import org.hibernate.annotations.GenericGenerator;
-import org.hibernate.engine.jdbc.BlobProxy;
 
 @Entity
 public class WatchEntity extends BaseEntity<Long> {
@@ -20,7 +16,6 @@ public class WatchEntity extends BaseEntity<Long> {
 	private static final long serialVersionUID = 1L;
 
 	@Id
-    //@GeneratedValue(strategy = GenerationType.AUTO)
 	@GeneratedValue(generator="increment")
 	@GenericGenerator(name="increment", strategy = "increment")
     @Column(name="id")
@@ -30,7 +25,7 @@ public class WatchEntity extends BaseEntity<Long> {
     private String title;
 
 	@Column(name="price", nullable=false)
-    private int price;
+    private BigDecimal price;
 
 	@Column(name="description", nullable=true, length=2048)
     private String description;
@@ -67,11 +62,11 @@ public class WatchEntity extends BaseEntity<Long> {
 		return this;
 	}
 
-	public int getPrice() {
+	public BigDecimal getPrice() {
 		return price;
 	}
 
-	public WatchEntity setPrice(int price) {
+	public WatchEntity setPrice(BigDecimal price) {
 		this.price = price;
 		return this;
 	}
@@ -94,34 +89,5 @@ public class WatchEntity extends BaseEntity<Long> {
 	  }
 	 
 
-	
-	
-	
-	/*
-	 * public String getFountainString() { return fountain;
-	 * 
-	 * if(fountain != null ) { try { return
-	 * (Base64.getEncoder().encodeToString(fountain.getBytes(0,
-	 * (int)fountain.length()))); } catch (SQLException e) { throw new
-	 * RuntimeException(e); } } else { return fountainString; }
-	 * 
-	 * }
-	 * 
-	 * public void setFountainString(String fountainString) { this.fountain =
-	 * fountainString;
-	 * 
-	 * byte[] decodedBytes = Base64.getDecoder().decode(fountainString); fountain =
-	 * BlobProxy.generateProxy(decodedBytes);
-	 * 
-	 * }
-	 */
-	
-	/*
-	 * public Blob getFountain() { return fountain; }
-	 * 
-	 * public WatchEntity setFountain(Blob fountain) { this.fountain = fountain;
-	 * return this; }
-	 */	 
-	
 	
 }
